+++ src/web/Controller.php
@@ -9,6 +9,7 @@ namespace craft\web;
 
 use Craft;
 use craft\base\ModelInterface;
+use craft\elements\User;
 use yii\base\Action;
 use yii\base\InvalidArgumentException;
 use yii\base\InvalidConfigException;
@@ -173,6 +174,19 @@ abstract class Controller extends \yii\web\Controller
         }
 
         return true;
+    }
+
+    /**
+     * Returns the currently logged-in user.
+     *
+     * @param bool $autoRenew
+     * @return ?User
+     * @see \yii\web\User::getIdentity()
+     * @since 4.3.0
+     */
+    public static function currentUser(bool $autoRenew = true): ?User
+    {
+        return Craft::$app->getUser()->getIdentity($autoRenew);
     }
 
     /**